+++ src/index.tsx
@@ -24,71 +24,9 @@ import {
 } from 'lodash';
 
 import 'vis-network/styles/vis-network.css';
+import { GraphEvents } from './EventTypes';
 
 export type { Network, Edge, Node, Options, NetworkEvents, IdType };
-
-export type GraphEvents = Partial<
-  // eslint-disable-next-line @typescript-eslint/no-explicit-any
-   Record<UnsupportedTypeNetworkEvents, (params?: any) => void>
-  | Record<'click', (params: SingleClickData) => void>
-  | Record<
-      | 'doubleClick'
-      | 'oncontext'
-      | 'hold'
-      | 'release'
-      | 'select'
-      | 'selectNode'
-      | 'selectEdge'
-      | 'dragStart'
-      | 'dragging'
-      | 'dragEnd',
-      (params: BaseClickData) => void
-    >
->;
-
-type UnsupportedTypeNetworkEvents =
-  | 'deselectNode'
-  | 'deselectEdge'
-  | 'controlNodeDragging'
-  | 'controlNodeDragEnd'
-  | 'hoverNode'
-  | 'blurNode'
-  | 'hoverEdge'
-  | 'blurEdge'
-  | 'zoom'
-  | 'showPopup'
-  | 'hidePopup'
-  | 'startStabilizing'
-  | 'stabilizationProgress'
-  | 'stabilizationIterationsDone'
-  | 'stabilized'
-  | 'resize'
-  | 'initRedraw'
-  | 'beforeDrawing'
-  | 'afterDrawing'
-  | 'animationFinished'
-  | 'configChange';
-
-export interface BaseClickData {
-  nodes: string[];
-  edges: string[];
-  event: unknown; // TODO add event type
-  pointer: Pointer;
-}
-
-export interface SingleClickData extends BaseClickData {
-  items: unknown[]; // TODO add items type
-}
-
-export interface Pointer {
-  DOM: Position;
-  canvas: Position;
-}
-
-export interface Position {
-  x: number;
-  y: number;
-}
 
 export interface GraphData {
   nodes: Node[];
